+++ sdk/src/androidTest/java/ly/count/android/sdk/EventQueueTests.java
@@ -60,10 +60,10 @@ public class EventQueueTests extends AndroidTestCase {
         final double sum = 3.0d;
         final double dur = 10.0d;
         final Map<String, String> segmentation = new HashMap<String, String>(1);
-        final int timestamp = (int)Countly.currentTimestampMs();
+        final long timestamp = Countly.currentTimestampMs();
         final int hour = Countly.currentHour();
         final int dow = Countly.currentDayOfWeek();
-        final ArgumentCaptor<Integer> arg = ArgumentCaptor.forClass(Integer.class);
+        final ArgumentCaptor<Long> arg = ArgumentCaptor.forClass(Long.class);
 
         mEventQueue.recordEvent(eventKey, segmentation, count, sum, dur);
         verify(mMockCountlyStore).addEvent(eq(eventKey), eq(segmentation), arg.capture(), eq(hour), eq(dow), eq(count), eq(sum), eq(dur));
